# server/openapi/swagger.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Traveling Cook Calculator API
  description: |
    API for managing Cook and Run events - a social cooking experience where teams 
    cook different courses at different locations and move between them.

    ## Authentication
    All endpoints require a valid JWT Bearer token. The JWT should contain:
    - `sub`: User ID
    - `role`: User role (admin, organizer, participant)
    - `exp`: Token expiration time

    ## Rate Limiting
    - 1000 requests per hour per authenticated user
    - 100 requests per hour for unauthenticated endpoints

    ## Error Handling
    The API uses conventional HTTP response codes and returns error details in JSON format.
    All error responses follow the same structure with error codes, messages, and optional details.
servers:
  - url: http://0.0.0.0:3000/
    description: Local development server
security:
  - bearerAuth: []
paths:
...
  /cook_and_run/{cookAndRunId}/teams:
    get:
      summary: List all teams for a cook and run project
      description: Returns all teams with optional filtering, sorting and pagination
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [name_asc, name_desc, created_asc, created_desc]
            default: created_asc
      responses:
        "200":
          description: Paginated list of teams
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Team"
                  pagination:
                    $ref: "#/components/schemas/PaginationInfo"
                required:
                  - data
                  - pagination
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Teams
  /cook_and_run/{cookAndRunId}/team/{teamId}:
    post:
      summary: Create team for Cook and Run project
      description: Creates a new team for the specified cook and run project
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
        - $ref: "#/components/parameters/TeamId"
      requestBody:
        $ref: "#/components/requestBodies/CookAndRunCreateTeamRequest"
      responses:
        "201":
          description: Team created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "409":
          description: Team with this ID already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Teams

    get:
      summary: Get team details
      description: Returns detailed information about a specific team
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
        - $ref: "#/components/parameters/TeamId"
      responses:
        "200":
          description: Team details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Teams

    patch:
      summary: Update team for Cook and Run project
      description: Updates an existing team with new data
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
        - $ref: "#/components/parameters/TeamId"
      requestBody:
        $ref: "#/components/requestBodies/CookAndRunUpdateTeamRequest"
      responses:
        "200":
          description: Team updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Teams

    delete:
      summary: Delete team for cook and run project
      description: Permanently deletes a team from the cook and run project
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
        - $ref: "#/components/parameters/TeamId"
      responses:
        "204":
          description: Team deleted successfully
        "400":
          description: Cannot delete team that is assigned to host courses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Teams
  /cook_and_run/{cookAndRunId}/team/{teamId}/plan:
    get:
      summary: Get personalized plan for team
      description: |
        Returns the personalized walking plan for a specific team, including
        their hosting assignments and route between other teams' locations
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
        - $ref: "#/components/parameters/TeamId"
      responses:
        "200":
          $ref: "#/components/responses/CookAndRunPlanResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Teams
...
  /cook_and_run/{cookAndRunId}/share_team_config:
    patch:
      summary: Create team sharing configuration
      description: |
        Creates a configuration for public team registration.
        This allows teams to self-register for the event via a shared link.
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
      requestBody:
        $ref: "#/components/requestBodies/CookAndRunTeamShareCreateRequest"
      responses:
        "201":
          description: Share configuration created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ShareTeamConfig"
                  - type: object
                    properties:
                      share_url:
                        type: string
                        format: uri
                        description: Public URL for team registration
                        example: "https://app.travelingcook.com/register/abc123"
        "409":
          description: Share configuration already exists for this project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Team Sharing

    get:
      summary: Get team sharing configuration
      description: Returns the current team sharing configuration
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
      responses:
        "200":
          description: Share configuration details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ShareTeamConfig"
                  - type: object
                    properties:
                      share_url:
                        type: string
                        format: uri
                        description: Public URL for team registration
                      registration_count:
                        type: integer
                        description: Number of teams registered via this link
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Team Sharing

    delete:
      summary: Delete team sharing configuration
      description: |
        Disables public team registration by deleting the share configuration.
        Existing shared links will no longer work.
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
      responses:
        "204":
          description: Share configuration deleted successfully
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Team Sharing
...
components:
  parameters:
    CookAndRunId:
      name: cookAndRunId
      in: path
      required: true
      description: Unique identifier for the cook and run project
      schema:
        type: string
        format: uuid
      example: "123e4567-e89b-12d3-a456-426614174000"
...
    TeamId:
      name: teamId
      in: path
      required: true
      description: Unique identifier for the team
      schema:
        type: string
        format: uuid
      example: "789e0123-e45f-67g8-h901-234567890123"
...
  requestBodies:
...
    CookAndRunCreateTeamRequest:
      description: Request body to create team for a Cook and Run project
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TeamCreateData"
          examples:
            basic_team:
              summary: Basic team registration
              value:
                name: "The Cooking Ninjas"
                userId: "user456"
                address:
                  address: "Maximilianstrasse 10, 80539 München, Germany"
                  latitude: 48.142385
                  longitude: 11.579381
                needs_check: true
            detailed_team:
              summary: Team with full contact details
              value:
                name: "Gourmet Wanderers"
                userId: "user789"
                address:
                  address: "Leopoldstrasse 50, 80802 München, Germany"
                  latitude: 48.154851
                  longitude: 11.583903
                mail: "gourmets@example.com"
                phone: "+49 89 123456789"
                members: 4
                diets: "2x vegetarian, 1x gluten-free"
                needs_check: true
    CookAndRunUpdateTeamRequest:
      description: Request body to update team for a Cook and Run project
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TeamUpdateData"
...
    CookAndRunTeamShareCreateRequest:
      description: Request body to create share config for Cook and Run teams
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - invite_text
              - needs_login
              - default_needs_check
              - required_fields
            properties:
              invite_text:
                type: string
                minLength: 1
                maxLength: 1000
                description: Welcome message shown to teams during registration
              needs_login:
                type: boolean
                description: Whether teams need to create user accounts
              default_needs_check:
                type: boolean
                description: Whether new team registrations need manual approval
              required_fields:
                type: array
                items:
                  type: string
                  enum: [mail, phone, members, diets]
                minItems: 0
                maxItems: 4
                uniqueItems: true
                description: Fields that teams must fill out during registration
              max_teams:
                type: integer
                minimum: 1
                maximum: 100
                description: Maximum number of teams that can register
              registration_deadline:
                type: string
                format: date-time
                description: Deadline for team registration
          examples:
            open_registration:
              summary: Open registration with minimal requirements
              value:
                invite_text: "Join our amazing Cook & Run event! Register your team and get ready for a culinary adventure."
                needs_login: false
                default_needs_check: false
                required_fields: ["mail"]
            controlled_registration:
              summary: Controlled registration with approval
              value:
                invite_text: "Welcome to our exclusive Cook & Run event. All team registrations will be reviewed before approval."
                needs_login: true
                default_needs_check: true
                required_fields: ["mail", "phone", "members", "diets"]
                max_teams: 20
                registration_deadline: "2024-06-01T23:59:59Z"
...
  responses:
    BadRequestError:
      description: Bad request - invalid input parameters or request format
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalid_uuid:
              summary: Invalid UUID format
              value:
                error: "BAD_REQUEST"
                message: "Invalid UUID format in path parameter"
                timestamp: "2024-03-15T10:00:00Z"
            missing_field:
              summary: Missing required field
              value:
                error: "BAD_REQUEST"
                message: "Missing required fields in request body"
                details:
                  - field: "name"
                    message: "Name is required and cannot be empty"
                timestamp: "2024-03-15T10:00:00Z"
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missing_token:
              summary: Missing authentication token
              value:
                error: "UNAUTHORIZED"
                message: "Authentication token is required"
                timestamp: "2024-03-15T10:00:00Z"
            invalid_token:
              summary: Invalid or expired token
              value:
                error: "UNAUTHORIZED"
                message: "Authentication token is invalid or expired"
                timestamp: "2024-03-15T10:00:00Z"
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            project_not_found:
              summary: Cook and Run project not found
              value:
                error: "NOT_FOUND"
                message: "Cook and Run project not found"
                timestamp: "2024-03-15T10:00:00Z"
            team_not_found:
              summary: Team not found
              value:
                error: "NOT_FOUND"
                message: "Team not found in this project"
                timestamp: "2024-03-15T10:00:00Z"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validation_failed:
              summary: Multiple validation errors
              value:
                error: "VALIDATION_ERROR"
                message: "Request validation failed"
                details:
                  - field: "email"
                    message: "Invalid email format"
                  - field: "phone"
                    message: "Phone number must be in international format"
                  - field: "members"
                    message: "Number of members must be between 1 and 20"
                timestamp: "2024-03-15T10:00:00Z"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            server_error:
              summary: Unexpected server error
              value:
                error: "INTERNAL_SERVER_ERROR"
                message: "An unexpected error occurred. Please try again later."
                timestamp: "2024-03-15T10:00:00Z"
...
    CookAndRunPlanResponse:
      description: Personalized plan for a team
      content:
        application/json:
          schema:
            type: object
            properties:
              introduction:
                type: string
                maxLength: 1000
                description: Welcome text and instructions for participants
              current_team:
                $ref: "#/components/schemas/TeamSummary"
              walking_path:
                type: array
                items:
                  type: object
                  properties:
                    course:
                      $ref: "#/components/schemas/Course"
                    team:
                      $ref: "#/components/schemas/TeamSummary"
          examples:
            team_plan:
              summary: Complete walking plan for a team
              value:
                introduction: "Welcome to our Cook & Run event! Follow your personalized plan below."
                current_team:
                  id: "789e0123-e45f-67g8-h901-234567890123"
                  name: "The Cooking Ninjas"
                  address:
                    address: "Maximilianstrasse 10, 80539 München, Germany"
                    latitude: 48.142385
                    longitude: 11.579381
                walking_path:
                  - course:
                      id: "app-001"
                      name: "Appetizers"
                      time: "18:00"
                    team:
                      id: "team-host-1"
                      name: "Gourmet Starters"
                      address:
                        address: "Leopoldstrasse 50, 80802 München"
                        latitude: 48.154851
                        longitude: 11.583903
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Validation failed for one or more fields"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Field that caused the error
              message:
                type: string
                description: Detailed error message for this field
            required:
              - field
              - message
          description: Detailed error information (optional)
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
        trace_id:
          type: string
          description: Unique identifier for tracing this error (optional)
      required:
        - error
        - message
        - timestamp
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
          example: 1
        limit:
          type: integer
          minimum: 1
          description: Number of items per page
          example: 20
        total:
          type: integer
          minimum: 0
          description: Total number of items across all pages
          example: 45
        total_pages:
          type: integer
          minimum: 0
          description: Total number of pages
          example: 3
        has_next:
          type: boolean
          description: Whether there are more pages after this one
          example: true
        has_prev:
          type: boolean
          description: Whether there are pages before this one
          example: false
      required:
        - page
        - limit
        - total
        - total_pages
        - has_next
        - has_prev
    Address:
      type: object
      properties:
        address:
          type: string
          minLength: 1
          maxLength: 256
          description: Human-readable address
          example: "Marienplatz 1, 80331 München, Germany"
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
          description: Latitude coordinate (WGS84)
          example: 48.137154
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
          description: Longitude coordinate (WGS84)
          example: 11.575490
      required:
        - address
        - latitude
        - longitude
...
    Team:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the team
        created_by_user:
          type: string
          maxLength: 64
          description: ID of the user who created this team
        name:
          type: string
          maxLength: 64
          description: Team name
        created:
          type: string
          format: date-time
          description: When the team was created
        edited:
          type: string
          format: date-time
          description: When the team was last modified
        address:
          $ref: "#/components/schemas/Address"
        mail:
          type: string
          format: email
          maxLength: 254
          description: Contact email for the team
        phone:
          type: string
          pattern: '^[\+]?[1-9][\d\s\-\(\)]{7,15}'
          maxLength: 20
          description: International phone number
        members:
          type: integer
          minimum: 1
          maximum: 20
          description: Number of team members
        diets:
          type: string
          maxLength: 256
          description: Special dietary requirements and allergies
        needs_check:
          type: boolean
          description: Whether team registration needs manual approval
        note_list:
          type: array
          items:
            $ref: "#/components/schemas/Note"
          description: Notes associated with this team
      required:
        - id
        - name
        - created
        - edited
        - address
        - needs_check
        - status
        - note_list
...
    Note:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the note
        headline:
          type: string
          maxLength: 64
          description: Short headline for the note
        content:
          type: string
          maxLength: 1000
          description: Detailed content of the note
        created:
          type: string
          format: date-time
          description: When the note was created
      required:
        - id
        - headline
        - content
        - created
    ShareTeamConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the share configuration
        invite_text:
          type: string
          maxLength: 1000
          description: Welcome message shown during registration
        needs_login:
          type: boolean
          description: Whether teams need to create user accounts
        default_needs_check:
          type: boolean
          description: Whether new registrations need manual approval
        required_fields:
          type: array
          items:
            type: string
            enum: [mail, phone, members, diets]
          minItems: 0
          maxItems: 4
          uniqueItems: true
          description: Fields that teams must fill out during registration
        max_teams:
          type: integer
          minimum: 1
          maximum: 100
          description: Maximum number of teams that can register
        registration_deadline:
          type: string
          format: date-time
          description: Deadline for team registration
        created:
          type: string
          format: date-time
          description: When the share configuration was created
      required:
        - id
        - invite_text
        - needs_login
        - default_needs_check
        - required_fields
        - created
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token containing user authentication information.

        Token should include the following claims:
        - `sub`: User ID
        - `role`: User role (admin, organizer, participant)
        - `exp`: Token expiration time
        - `iss`: Token issuer (api.travelingcook.com)

        Format: `Authorization: Bearer <token>`

        Example: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`

# Security considerations:
# - Tokens should be refreshed every 24 hours
# - Use HTTPS in production
# - Implement proper token validation and blacklisting
# - Rate limiting is applied per authenticated user
...